// app/frontend/src/types.ts
export interface Listing {
  id: string;
  title: string;
  description: string;
  price: string;
  rooms: string;
  area: string;
  location: string;
  district: string;
  zipCode: string;
  floor: string;
  yearBuilt: string;
  condition: string;
  pictureCount: number;
  imageUrl: string;
  exposeUrl: string;
  type: string;
  balcony: boolean;
  garden: boolean;
  cellar: boolean;
  lift: boolean;
  parking: boolean;
}